pyproject.toml
[tool.poetry]
name = "icnft-script"
version = "0.1.0"
description = ""
authors = ["Oda <oda@xyzzyx.co.jp>"]

[tool.poetry.dependencies]
python = "^3.10.4"
ic-py = "^1.0.0"
python-dotenv = "^0.20.0"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

